Filter message edits
Allows Chrysanthemum to respond to message edits. This required some
extra work to respect the cache, which should allow Chrysanthemum to be
more performant in the critical path of a user sending a message and
then immediately editing it.

This change will increase Chrysanthemum's memory usage significantly,
due to the cache now containing member and user information. We'll have
to keep an eye on this. We might have to build our own bespoke cache
that we can clear out more frequently if the added weight of user and
member information is too high. We'll see what it looks like in
deployment.

Closes #11.