include filter name in test command output
Also redid the logic for /chrysanthemum-test so that it short-circuits if one filter fails on a message. This may visibly change the output for a given test, depending on the filter configuration, but it should match the behavior of the actual filtering logic, which will select the first failing filter.

Closes #33