Redo command handling again
We no longer have the really esoteric logic for tracking command IDs and instead just filter by the command name that Discord spits back out. This dramatically simplifies the amount of state we're keeping.

To pull this off, we switch to using set_guild_commands for each guild, which will overwrite the command list. This has the beneficial side effect of cleaning up old commands if we ever change around the command set again.